Replace the AI provider if/else chains with a provider registry, split aiService
aiService.js was a 964-line module owning provider dispatch (a five-way
if/else duplicated across the streaming and non-streaming paths),
Discord transport, rate limiting, RAG, action execution, pricing and
usage accounting. Adding a provider meant editing four separate sites.

Define one provider interface — { name, label, defaultModel, pricing,
resolveAuth, validateModel?, stream(req), complete(req) } — implemented
by ai/providers/{openai,gemini,anthropic,ollama,openrouter}.js and
registered in a map in ai/providers/index.js. Both dispatch paths are
now a single registry lookup, per-provider pricing tables live on the
provider, default models and Discord-facing labels derive from the
registry, and the OpenRouter model-prefix check moved behind the
interface as validateModel.

The rest of the module splits along the lines it already had:
ai/usage.js (cost + ledger), ai/knowledge.js (RAG), ai/history.js,
ai/rateLimit.js, ai/actions.js, and a thin ai/discordChat.js owning the
transport. services/aiService.js stays as a façade with the exact same
exports, so no import site or test changes.

Closes #615
